[SHIP-1499] Fix a bug with expanding grouped rows
In a previous PR I made it so that the group would have the column *name* as
a label, rather than the column key. However, this same prop was also passed to
the event handler on the consumer, which broke things as it no longer matched.

Now a seperate prop containing the key is used for the event handler.